rawnutrition/requirements.txt
django-widget-tweaks == 1.5.0